OLDNEW for LA114.wind; which switches to a different strategy for evaluating the country level wind supply curves (we wanted to do so for a couple of reasons: avoid excessive loops, and vastly speed up the calculation).  The OLD_DATA_SYSTEM just read the previous price points from temp-data-inject because the old method was so slow.
The old approach was to continue to evaluate the curves until the value
wasn't changing anymore with in some tolerance (with some minumum number
of calculations to ensure we don't quit in a shallow start to the
curve).  The new approach would only evaluate prices evenly spaces from
min.price to the 95th percentile of max production (the percent is
configurable)

Note this removed the need for
temp-data-inject/L114.OLDregional_price_supply_points.csv the last
temp-data!